+++ backend/schedulerd/state_manager.go
@@ -1,6 +1,7 @@
 package schedulerd
 
 import (
+	"context"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -35,17 +36,21 @@ func NewStateManager(store store.Store) *StateManager {
 
 	manager.synchronizer = NewSynchronizeStateScheduler(
 		SynchronizeMinInterval,
-		&SyncronizeChecks{
+		&SynchronizeChecks{
 			Store:    store,
 			OnUpdate: manager.updateChecks,
 		},
-		&SyncronizeAssets{
+		&SynchronizeAssets{
 			Store:    store,
 			OnUpdate: manager.updateAssets,
 		},
-		&SyncronizeHooks{
+		&SynchronizeHooks{
 			Store:    store,
 			OnUpdate: manager.updateHooks,
+		},
+		&SynchronizeEntities{
+			Store:    store,
+			OnUpdate: manager.updateEntities,
 		},
 	)
 
@@ -53,8 +58,8 @@ func NewStateManager(store store.Store) *StateManager {
 }
 
 // Start keeping state synchronized
-func (mngrPtr *StateManager) Start() {
-	mngrPtr.synchronizer.Start()
+func (mngrPtr *StateManager) Start(ctx context.Context) {
+	mngrPtr.synchronizer.Start(ctx)
 }
 
 // Stop keeping state synchronized
@@ -101,6 +106,12 @@ func (mngrPtr *StateManager) updateHooks(hooks []*types.HookConfig) {
 	})
 }
 
+func (mngrPtr *StateManager) updateEntities(entities []*types.Entity) {
+	mngrPtr.updateState(func(state *SchedulerState) {
+		state.SetEntities(entities)
+	})
+}
+
 func (mngrPtr *StateManager) updateState(updateFn func(newState *SchedulerState)) {
 	// Lock to avoid competing updates
 	mngrPtr.mutex.Lock()
@@ -134,9 +145,10 @@ func (mngrPtr *StateManager) updateSyncInterval() {
 
 // A SchedulerState represents the internal state of the cache
 type SchedulerState struct {
-	checks map[string]*types.CheckConfig
-	assets map[string]map[string]*types.Asset
-	hooks  map[string]map[string]*types.HookConfig
+	checks   map[string]*types.CheckConfig
+	assets   map[string]map[string]*types.Asset
+	hooks    map[string]map[string]*types.HookConfig
+	entities map[string]map[string]*types.Entity
 }
 
 // GetCheck returns check given name and organization
@@ -145,18 +157,30 @@ func (statePtr *SchedulerState) GetCheck(name, org, env string) *types.CheckConf
 	return statePtr.checks[key]
 }
 
-// GetAssetsInOrg returns all assets associated given organization
-func (statePtr *SchedulerState) GetAssetsInOrg(org string) (res []*types.Asset) {
+// GetAssetsInNamespace returns all assets associated given organization
+func (statePtr *SchedulerState) GetAssetsInNamespace(org string) (res []*types.Asset) {
 	for _, asset := range statePtr.assets[org] {
 		res = append(res, asset)
 	}
 	return
 }
 
-// GetHooksInOrg returns all hooks associated given organization
-func (statePtr *SchedulerState) GetHooksInOrg(org string) (res []*types.HookConfig) {
-	for _, hook := range statePtr.hooks[org] {
+// GetHooksInNamespace returns all hooks associated given organization
+// and environment
+func (statePtr *SchedulerState) GetHooksInNamespace(org string, env string) (res []*types.HookConfig) {
+	orgEnv := concatUniqueKey(org, env)
+	for _, hook := range statePtr.hooks[orgEnv] {
 		res = append(res, hook)
+	}
+	return
+}
+
+// GetEntitiesInNamespace returns all entities associated given organization
+// and environment
+func (statePtr *SchedulerState) GetEntitiesInNamespace(org string, env string) (res []*types.Entity) {
+	orgEnv := concatUniqueKey(org, env)
+	for _, entity := range statePtr.entities[orgEnv] {
+		res = append(res, entity)
 	}
 	return
 }
@@ -185,6 +209,14 @@ func (statePtr *SchedulerState) SetHooks(hooks []*types.HookConfig) {
 	}
 }
 
+// SetEntities overwrites current set of entities w/ given
+func (statePtr *SchedulerState) SetEntities(entities []*types.Entity) {
+	statePtr.entities = make(map[string]map[string]*types.Entity)
+	for _, entity := range entities {
+		statePtr.addEntity(entity)
+	}
+}
+
 func (statePtr *SchedulerState) addCheck(check *types.CheckConfig) {
 	key := concatUniqueKey(check.Name, check.Organization, check.Environment)
 	statePtr.checks[key] = check
@@ -202,12 +234,26 @@ func (statePtr *SchedulerState) addAsset(asset *types.Asset) {
 
 func (statePtr *SchedulerState) addHook(hook *types.HookConfig) {
 	org := hook.Organization
-	if orgMap := statePtr.hooks[org]; orgMap == nil {
-		statePtr.hooks[org] = make(map[string]*types.HookConfig)
+	env := hook.Environment
+	orgEnv := concatUniqueKey(org, env)
+	if hookMap := statePtr.hooks[orgEnv]; hookMap == nil {
+		statePtr.hooks[orgEnv] = make(map[string]*types.HookConfig)
 	}
 
-	key := concatUniqueKey(hook.Name, org)
-	statePtr.hooks[org][key] = hook
+	key := concatUniqueKey(hook.Name, org, env)
+	statePtr.hooks[orgEnv][key] = hook
+}
+
+func (statePtr *SchedulerState) addEntity(entity *types.Entity) {
+	org := entity.Organization
+	env := entity.Environment
+	orgEnv := concatUniqueKey(org, env)
+	if entityMap := statePtr.entities[orgEnv]; entityMap == nil {
+		statePtr.entities[orgEnv] = make(map[string]*types.Entity)
+	}
+
+	key := concatUniqueKey(entity.ID, org, env)
+	statePtr.entities[orgEnv][key] = entity
 }
 
 func concatUniqueKey(args ...string) string {