+++ CHANGELOG.md
@@ -8,12 +8,17 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
 ### Added
 - Add function for matching entities to a proxy check request.
-
-### Changed
-- Govaluate logic is now wrapped in the `util/eval` package.
 - Added functions for publishing proxy check requests.
 - Added proxy request validation.
 - CLI functionality for proxy check requests (add set-proxy-requests command).
+- Entities have been added to the state manager and synchronizer.
+
+### Changed
+- Govaluate logic is now wrapped in the `util/eval` package.
+
+### Fixed
+- Fixed a bug where retrieving check hooks were only from the check's
+organization, rather than the check's environment, too.
 
 ## [2.0.0-alpha.14] - 2018-01-23
 ### Added